requirements.txt
# =============================================================================
# Full SEO Automation - Python Dependencies
# =============================================================================
# Install: pip install -r requirements.txt
# =============================================================================

# --- Core Framework ---
sqlalchemy>=2.0,<3.0
aiosqlite>=0.19.0
alembic>=1.13.0
pydantic>=2.5.0,<3.0
pydantic-settings>=2.1.0,<3.0

# --- CLI & Terminal UI ---
typer[all]>=0.9.0,<1.0
rich>=13.7.0,<14.0

# --- Dashboard ---
streamlit>=1.30.0,<2.0
plotly>=5.18.0,<6.0

# --- Scheduling ---
apscheduler>=3.10.4,<4.0

# --- AI / LLM ---
openai>=1.12.0,<2.0
google-generativeai>=0.4.0,<1.0
tiktoken>=0.6.0,<1.0

# --- Google APIs ---
google-auth>=2.27.0,<3.0
google-auth-oauthlib>=1.2.0,<2.0
google-auth-httplib2>=0.2.0,<1.0
google-api-python-client>=2.116.0,<3.0
google-analytics-data>=0.18.0,<1.0

# --- Web Scraping & Browser ---
playwright>=1.41.0,<2.0
beautifulsoup4>=4.12.0,<5.0
lxml>=5.1.0,<6.0
httpx>=0.26.0,<1.0
aiohttp>=3.9.0,<4.0

# --- NLP & Text Processing ---
spacy>=3.7.0,<4.0
textstat>=0.7.3,<1.0

# --- SEO & Trends ---
pytrends>=4.9.2,<5.0

# --- Data & Configuration ---
pyyaml>=6.0.1,<7.0
python-dotenv>=1.0.0,<2.0

# --- Utilities ---
python-slugify>=8.0.1,<9.0
markdown>=3.5.0,<4.0
jinja2>=3.1.3,<4.0
cachetools>=5.3.2,<6.0

# --- Quality & Grammar ---
language-tool-python>=2.7.1,<3.0

# --- Testing ---
pytest>=8.0.0,<9.0
pytest-asyncio>=0.23.0,<1.0
pytest-cov>=4.1.0,<6.0
